root/synapse_experiment/requirements.txt
numpy
shapely
radon
ruff
pytest
pytest-cov
pandas
pyyaml 